Got the results back better, but I disklike the structure.
I'm not sure what's going wrong with modules.  I'd like to be
able to pass them as optional objects, so that the novice
doesn't have to know the itertools/functools packages.

But my lack of Python knowledge is showing, as I'm struggling.

Oh well, the code is working as I intended, even though the functionality
is not as cleanly seperated as I'd like for testing purposes.