Removed the call to contextlib.
ModuleNotFoundError is not available in Python < 3.6, so
rather that using contextlib.suppress with a specific set
of modules, instead, just do a try/except.

This is fine, and when I had put the contextlib stuff in there
the try/except lines were far away from each other, therefore
suppress was a better idea.

Since the code is better factored now, a try/except replacement
for suppress is reasonable, especially given support for Python < 3.6.